requirements.txt
streamlit~=1.28.2
python-dotenv~=1.0.0
pypdf2~=3.0.1
langchain~=0.0.335
cohere